+++ crates/op-rbuilder/Cargo.toml
@@ -98,6 +98,7 @@ rand = "0.9.0"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "json"] }
 shellexpand = "3.1"
 serde_yaml = { version = "0.9" }
+dashmap = "6.1"
 
 
 # `msozin/flashblocks-v1.4.1` branch based on `flashblocks-rebase`
