+++ crates/op-rbuilder/src/builders/flashblocks/mod.rs
@@ -1,0 +1,42 @@
+use super::BuilderConfig;
+use crate::traits::{NodeBounds, PoolBounds};
+use config::FlashblocksConfig;
+use service::FlashblocksServiceBuilder;
+
+mod config;
+mod context;
+mod payload;
+mod service;
+mod wspub;
+
+/// Block building strategy that progressively builds chunks of a block and makes them available
+/// through a websocket update, then merges them into a full block every chain block time.
+pub struct FlashblocksBuilder;
+
+impl super::PayloadBuilder for FlashblocksBuilder {
+    type Config = FlashblocksConfig;
+
+    type ServiceBuilder<Node, Pool>
+        = FlashblocksServiceBuilder
+    where
+        Node: NodeBounds,
+        Pool: PoolBounds;
+
+    fn new_service<Node, Pool>(
+        config: BuilderConfig<Self::Config>,
+    ) -> eyre::Result<Self::ServiceBuilder<Node, Pool>>
+    where
+        Node: NodeBounds,
+        Pool: PoolBounds,
+    {
+        Ok(FlashblocksServiceBuilder(config))
+    }
+}
+
+#[derive(Debug, Default)]
+struct ExtraExecutionInfo {
+    /// Index of the last consumed flashblock
+    pub last_flashblock_index: usize,
+}
+
+type ExecutionInfo = crate::primitives::reth::ExecutionInfo<ExtraExecutionInfo>;
