+++ crates/op-rbuilder/src/builders/flashblocks/service.rs
@@ -1,0 +1,56 @@
+use super::{payload::OpPayloadBuilder, FlashblocksConfig};
+use crate::{
+    builders::{generator::BlockPayloadJobGenerator, BuilderConfig},
+    traits::{NodeBounds, PoolBounds},
+};
+use reth_basic_payload_builder::BasicPayloadJobGeneratorConfig;
+use reth_node_api::NodeTypes;
+use reth_node_builder::{components::PayloadServiceBuilder, BuilderContext};
+use reth_optimism_evm::OpEvmConfig;
+use reth_payload_builder::{PayloadBuilderHandle, PayloadBuilderService};
+use reth_provider::CanonStateSubscriptions;
+
+pub struct FlashblocksServiceBuilder(pub BuilderConfig<FlashblocksConfig>);
+
+impl<Node, Pool> PayloadServiceBuilder<Node, Pool, OpEvmConfig> for FlashblocksServiceBuilder
+where
+    Node: NodeBounds,
+    Pool: PoolBounds,
+{
+    async fn spawn_payload_builder_service(
+        self,
+        ctx: &BuilderContext<Node>,
+        pool: Pool,
+        _: OpEvmConfig,
+    ) -> eyre::Result<PayloadBuilderHandle<<Node::Types as NodeTypes>::Payload>> {
+        tracing::debug!("Spawning flashblocks payload builder service");
+
+        let payload_builder = OpPayloadBuilder::new(
+            OpEvmConfig::optimism(ctx.chain_spec()),
+            pool,
+            ctx.provider().clone(),
+            self.0.clone(),
+        )?;
+
+        let payload_job_config = BasicPayloadJobGeneratorConfig::default();
+
+        let payload_generator = BlockPayloadJobGenerator::with_builder(
+            ctx.provider().clone(),
+            ctx.task_executor().clone(),
+            payload_job_config,
+            payload_builder,
+            true,
+            self.0.block_time_leeway,
+        );
+
+        let (payload_service, payload_builder) =
+            PayloadBuilderService::new(payload_generator, ctx.provider().canonical_state_stream());
+
+        ctx.task_executor()
+            .spawn_critical("custom payload builder service", Box::pin(payload_service));
+
+        tracing::info!("Flashblocks payload builder service started");
+
+        Ok(payload_builder)
+    }
+}
