+++ crates/op-rbuilder/src/builders/flashblocks/wspub.rs
@@ -1,0 +1,226 @@
+use core::{
+    fmt::{Debug, Formatter},
+    net::SocketAddr,
+    pin::Pin,
+    sync::atomic::{AtomicUsize, Ordering},
+    task::{Context, Poll},
+};
+use futures::{Sink, SinkExt};
+use rollup_boost::primitives::FlashblocksPayloadV1;
+use std::sync::Arc;
+use std::{io, net::TcpListener};
+use tokio::{
+    net::TcpStream,
+    sync::{
+        broadcast::{self, error::RecvError, Receiver},
+        watch,
+    },
+};
+use tokio_tungstenite::{
+    accept_async,
+    tungstenite::{Message, Utf8Bytes},
+    WebSocketStream,
+};
+use tracing::warn;
+
+/// A WebSockets publisher that accepts connections from client websockets and broadcasts to them
+/// updates about new flashblocks. It maintains a count of sent messages and active subscriptions.
+///
+/// This is modelled as a `futures::Sink` that can be used to send `FlashblocksPayloadV1` messages.
+pub struct WebSocketPublisher {
+    sent: Arc<AtomicUsize>,
+    subs: Arc<AtomicUsize>,
+    term: watch::Sender<bool>,
+    pipe: broadcast::Sender<Utf8Bytes>,
+}
+
+impl WebSocketPublisher {
+    pub fn new(addr: SocketAddr) -> io::Result<Self> {
+        let (pipe, _) = broadcast::channel(100);
+        let (term, _) = watch::channel(false);
+
+        let sent = Arc::new(AtomicUsize::new(0));
+        let subs = Arc::new(AtomicUsize::new(0));
+        let listener = TcpListener::bind(addr)?;
+
+        tokio::spawn(listener_loop(
+            listener,
+            pipe.subscribe(),
+            term.subscribe(),
+            Arc::clone(&sent),
+            Arc::clone(&subs),
+        ));
+
+        Ok(Self {
+            sent,
+            subs,
+            term,
+            pipe,
+        })
+    }
+
+    pub fn publish(&self, payload: &FlashblocksPayloadV1) -> io::Result<()> {
+        // Serialize the payload to a UTF-8 string
+        // serialize only once, then just copy around only a pointer
+        // to the serialized data for each subscription.
+        let serialized = serde_json::to_string(payload)?;
+        let utf8_bytes = Utf8Bytes::from(serialized);
+
+        // Send the serialized payload to all subscribers
+        self.pipe
+            .send(utf8_bytes)
+            .map_err(|e| io::Error::new(io::ErrorKind::ConnectionAborted, e))?;
+        Ok(())
+    }
+}
+
+impl Drop for WebSocketPublisher {
+    fn drop(&mut self) {
+        // Notify the listener loop to terminate
+        let _ = self.term.send(true);
+        tracing::info!("WebSocketPublisher dropped, terminating listener loop");
+    }
+}
+
+async fn listener_loop(
+    listener: TcpListener,
+    receiver: Receiver<Utf8Bytes>,
+    term: watch::Receiver<bool>,
+    sent: Arc<AtomicUsize>,
+    subs: Arc<AtomicUsize>,
+) {
+    let listener = tokio::net::TcpListener::from_std(listener)
+        .expect("Failed to convert TcpListener to tokio TcpListener");
+
+    let listen_addr = listener
+        .local_addr()
+        .expect("Failed to get local address of listener");
+    tracing::info!("Flashblocks WebSocketPublisher listening on {listen_addr}");
+
+    let mut term = term;
+
+    loop {
+        let subs = Arc::clone(&subs);
+
+        tokio::select! {
+            // drop this connection if the `WebSocketPublisher` is dropped
+            _ = term.changed() => {
+                if *term.borrow() {
+                    return;
+                }
+            }
+
+            // Accept new connections on the websocket listener
+            // when a new connection is established, spawn a dedicated task to handle
+            // the connection and broadcast with that connection.
+            Ok((connection, peer_addr)) = listener.accept() => {
+                let sent = Arc::clone(&sent);
+                let term = term.clone();
+                let receiver_clone = receiver.resubscribe();
+
+                match accept_async(connection).await {
+                    Ok(stream) => {
+                        tokio::spawn(async move {
+                            subs.fetch_add(1, Ordering::Relaxed);
+                            tracing::debug!("WebSocket connection established with {}", peer_addr);
+
+                            // Handle the WebSocket connection in a dedicated task
+                            broadcast_loop(stream, term, receiver_clone, sent).await;
+
+                            subs.fetch_sub(1, Ordering::Relaxed);
+                            tracing::debug!("WebSocket connection closed for {}", peer_addr);
+                        });
+                    }
+                    Err(e) => {
+                        warn!("Failed to accept WebSocket connection from {peer_addr}: {e}");
+                    }
+                }
+            }
+        }
+    }
+}
+
+/// An instance of this loop is spawned for each connected WebSocket client.
+/// It listens for broadcast updates about new flashblocks and sends them to the client.
+/// It also handles termination signals to gracefully close the connection.
+/// Any connectivity errors will terminate the loop, which will in turn
+/// decrement the subscription count in the `WebSocketPublisher`.
+async fn broadcast_loop(
+    stream: WebSocketStream<TcpStream>,
+    term: watch::Receiver<bool>,
+    blocks: broadcast::Receiver<Utf8Bytes>,
+    sent: Arc<AtomicUsize>,
+) {
+    let mut term = term;
+    let mut blocks = blocks;
+    let mut stream = stream;
+    let Ok(peer_addr) = stream.get_ref().peer_addr() else {
+        return;
+    };
+
+    loop {
+        tokio::select! {
+            // Check if the publisher is terminated
+            _ = term.changed() => {
+                if *term.borrow() {
+                    tracing::info!("WebSocketPublisher is terminating, closing broadcast loop");
+                    return;
+                }
+            }
+
+            // Receive payloads from the broadcast channel
+            payload = blocks.recv() => match payload {
+                Ok(payload) => {
+                    // Here you would typically send the payload to the WebSocket clients.
+                    // For this example, we just increment the sent counter.
+                    sent.fetch_add(1, Ordering::Relaxed);
+                    tracing::info!("Broadcasted payload: {:?}", payload);
+                    if let Err(e) = stream.send(Message::Text(payload)).await {
+                        tracing::debug!("Closing flashblocks subscription for {peer_addr}: {e}");
+                        break; // Exit the loop if sending fails
+                    }
+                }
+                Err(RecvError::Closed) => {
+                    tracing::debug!("Broadcast channel closed, exiting broadcast loop");
+                    return;
+                }
+                Err(RecvError::Lagged(_)) => {
+                    tracing::warn!("Broadcast channel lagged, some messages were dropped");
+                }
+            },
+        }
+    }
+}
+
+impl Debug for WebSocketPublisher {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+        let subs = self.subs.load(Ordering::Relaxed);
+        let sent = self.sent.load(Ordering::Relaxed);
+
+        f.write_str(&format!(
+            "WebSocketPublisher(subs={}, payloads={})",
+            subs, sent
+        ))
+    }
+}
+
+impl Sink<&FlashblocksPayloadV1> for WebSocketPublisher {
+    type Error = eyre::Report;
+
+    fn poll_ready(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn start_send(self: Pin<&mut Self>, item: &FlashblocksPayloadV1) -> Result<(), Self::Error> {
+        self.publish(item)?;
+        Ok(())
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn poll_close(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+}
