+++ crates/op-rbuilder/src/builders/mod.rs
@@ -1,5 +1,25 @@
+use core::{
+    convert::{Infallible, TryFrom},
+    fmt::Debug,
+    time::Duration,
+};
+use reth_node_builder::components::PayloadServiceBuilder;
+use reth_optimism_evm::OpEvmConfig;
+use reth_optimism_payload_builder::config::OpDAConfig;
+
+use crate::{
+    args::OpRbuilderArgs,
+    traits::{NodeBounds, PoolBounds},
+};
+
 mod flashblocks;
+mod generator;
 mod standard;
+
+mod old_standard;
+
+pub use flashblocks::FlashblocksBuilder;
+pub use standard::StandardBuilder;
 
 /// Defines the payload building mode for the OP builder.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
@@ -11,4 +31,121 @@ pub enum BuilderMode {
     /// block every short interval and makes it available through a websocket update
     /// then merges them into a full block every chain block time.
     Flashblocks,
+}
+
+/// Defines the interface for any block builder implementation API entry point.
+///
+/// Instances of this trait are used during Reth node construction as an argument
+/// to the `NodeBuilder::with_components` method to construct the payload builder
+/// service that gets called whenver the current node is asked to build a block.
+pub trait PayloadBuilder: Send + Sync + 'static {
+    /// The type that has an implementation specific variant of the Config<T> struct.
+    /// This is used to configure the payload builder service during startup.
+    type Config: TryFrom<OpRbuilderArgs, Error: Debug> + Clone + Debug + Send + Sync + 'static;
+
+    /// The type that is used to instantiate the payload builder service
+    /// that will be used by reth to build blocks whenever the node is
+    /// asked to do so.
+    type ServiceBuilder<Node, Pool>: PayloadServiceBuilder<Node, Pool, OpEvmConfig>
+    where
+        Node: NodeBounds,
+        Pool: PoolBounds;
+
+    /// Called during node startup by reth. Returns a [`PayloadBuilderService`] instance
+    /// that is preloaded with a [`PayloadJobGenerator`] instance specific to the builder
+    /// type.
+    fn new_service<Node, Pool>(
+        config: BuilderConfig<Self::Config>,
+    ) -> eyre::Result<Self::ServiceBuilder<Node, Pool>>
+    where
+        Node: NodeBounds,
+        Pool: PoolBounds;
+}
+
+/// Configuration values that are applicable to any type of block builder.
+#[derive(Clone)]
+pub struct BuilderConfig<Specific: Clone> {
+    /// When set to true, transactions are simulated by the builder and excluded from the block
+    /// if they revert. They may still be included in the block if individual transactions
+    /// opt-out of revert protection.
+    pub revert_protection: bool,
+
+    /// The interval at which blocks are added to the chain.
+    /// This is also the frequency at which the builder will be receiving FCU rquests from the
+    /// sequencer.
+    pub block_time: Duration,
+
+    /// Data Availability configuration for the OP builder
+    /// Defines constraints for the maximum size of data availability transactions.
+    pub da_config: OpDAConfig,
+
+    // The deadline is critical for payload availability. If we reach the deadline,
+    // the payload job stops and cannot be queried again. With tight deadlines close
+    // to the block number, we risk reaching the deadline before the node queries the payload.
+    //
+    // Adding 0.5 seconds as wiggle room since block times are shorter here.
+    // TODO: A better long-term solution would be to implement cancellation logic
+    // that cancels existing jobs when receiving new block building requests.
+    //
+    // When batcher's max channel duration is big enough (e.g. 10m), the
+    // sequencer would send an avalanche of FCUs/getBlockByNumber on
+    // each batcher update (with 10m channel it's ~800 FCUs at once).
+    // At such moment it can happen that the time b/w FCU and ensuing
+    // getPayload would be on the scale of ~2.5s. Therefore we should
+    // "remember" the payloads long enough to accommodate this corner-case
+    // (without it we are losing blocks). Postponing the deadline for 5s
+    // (not just 0.5s) because of that.
+    pub block_time_leeway: Duration,
+
+    /// Configuration values that are specific to the block builder implementation used.
+    pub specific: Specific,
+}
+
+impl<S: Debug + Clone> core::fmt::Debug for BuilderConfig<S> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        f.debug_struct("Config")
+            .field("revert_protection", &self.revert_protection)
+            .field("block_time", &self.block_time)
+            .field("block_time_leeway", &self.block_time_leeway)
+            .field("da_config", &self.da_config)
+            .field("specific", &self.specific)
+            .finish()
+    }
+}
+
+impl<S: Default + Clone> Default for BuilderConfig<S> {
+    fn default() -> Self {
+        Self {
+            revert_protection: false,
+            block_time: Duration::from_secs(2),
+            block_time_leeway: Duration::from_millis(500),
+            da_config: OpDAConfig::default(),
+            specific: S::default(),
+        }
+    }
+}
+
+impl<S> TryFrom<OpRbuilderArgs> for BuilderConfig<S>
+where
+    S: TryFrom<OpRbuilderArgs, Error: Debug> + Clone,
+{
+    type Error = S::Error;
+
+    fn try_from(args: OpRbuilderArgs) -> Result<Self, Self::Error> {
+        Ok(Self {
+            revert_protection: args.enable_revert_protection,
+            block_time: Duration::from_millis(args.chain_block_time),
+            block_time_leeway: Duration::from_millis(500),
+            da_config: Default::default(),
+            specific: S::try_from(args)?,
+        })
+    }
+}
+
+impl TryFrom<OpRbuilderArgs> for () {
+    type Error = Infallible;
+
+    fn try_from(_: OpRbuilderArgs) -> Result<Self, Self::Error> {
+        Ok(())
+    }
 }
