+++ crates/op-rbuilder/src/main.rs
@@ -1,6 +1,6 @@
-use args::CliExt;
-use clap::Parser;
-use reth_optimism_cli::{chainspec::OpChainSpecParser, Cli};
+use args::*;
+use builders::{BuilderConfig, BuilderMode, FlashblocksBuilder, StandardBuilder};
+use core::fmt::Debug;
 use reth_optimism_node::{
     node::{OpAddOnsBuilder, OpPoolBuilder},
     OpNode,
@@ -10,25 +10,19 @@ use reth_transaction_pool::TransactionPool;
 /// CLI argument parsing.
 pub mod args;
 mod builders;
-pub mod generator;
 mod metrics;
 mod monitor_tx_pool;
-#[cfg(feature = "flashblocks")]
-pub mod payload_builder;
 mod primitives;
 mod revert_protection;
 mod traits;
 mod tx;
 mod tx_signer;
 
-#[cfg(not(feature = "flashblocks"))]
-mod payload_builder_vanilla;
+// #[cfg(not(feature = "flashblocks"))]
+// mod payload_builder_vanilla;
 
-#[cfg(not(feature = "flashblocks"))]
-use payload_builder_vanilla::CustomOpPayloadBuilder;
-
-#[cfg(feature = "flashblocks")]
-use payload_builder::CustomOpPayloadBuilder;
+// #[cfg(not(feature = "flashblocks"))]
+// use payload_builder_vanilla::CustomOpPayloadBuilder;
 
 use metrics::{
     VersionInfo, BUILD_PROFILE_NAME, CARGO_PKG_VERSION, VERGEN_BUILD_TIMESTAMP,
@@ -44,85 +38,99 @@ use tx::FBPooledTransaction;
 static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
 fn main() {
-    let version = VersionInfo {
-        version: CARGO_PKG_VERSION,
-        build_timestamp: VERGEN_BUILD_TIMESTAMP,
-        cargo_features: VERGEN_CARGO_FEATURES,
-        git_sha: VERGEN_GIT_SHA,
-        target_triple: VERGEN_CARGO_TARGET_TRIPLE,
-        build_profile: BUILD_PROFILE_NAME,
+    let cli = Cli::parsed();
+    match cli.builder_mode() {
+        BuilderMode::Standard => {
+            tracing::info!("Starting OP builder in standard mode");
+            start_builder_node::<StandardBuilder>(cli);
+        }
+        BuilderMode::Flashblocks => {
+            tracing::info!("Starting OP builder in flashblocks mode");
+            start_builder_node::<FlashblocksBuilder>(cli);
+        }
     };
-
-    Cli::<OpChainSpecParser, args::OpRbuilderArgs>::parse()
-        .populate_defaults()
-        .run(|builder, builder_args| async move {
-            let rollup_args = builder_args.rollup_args;
-
-            let op_node = OpNode::new(rollup_args.clone());
-            let handle = builder
-                .with_types::<OpNode>()
-                .with_components(
-                    op_node
-                        .components()
-                        .pool(
-                            OpPoolBuilder::<FBPooledTransaction>::default()
-                                .with_enable_tx_conditional(
-                                    // Revert protection uses the same internal pool logic as conditional transactions
-                                    // to garbage collect transactions out of the bundle range.
-                                    rollup_args.enable_tx_conditional
-                                        || builder_args.enable_revert_protection,
-                                )
-                                .with_supervisor(
-                                    rollup_args.supervisor_http.clone(),
-                                    rollup_args.supervisor_safety_level,
-                                ),
-                        )
-                        .payload(CustomOpPayloadBuilder::new(
-                            builder_args.builder_signer,
-                            std::time::Duration::from_secs(builder_args.extra_block_deadline_secs),
-                            builder_args.flashblocks_ws_url,
-                            builder_args.chain_block_time,
-                            builder_args.flashblock_block_time,
-                        )),
-                )
-                .with_add_ons(
-                    OpAddOnsBuilder::default()
-                        .with_sequencer(rollup_args.sequencer.clone())
-                        .with_enable_tx_conditional(rollup_args.enable_tx_conditional)
-                        .build(),
-                )
-                .extend_rpc_modules(move |ctx| {
-                    if builder_args.enable_revert_protection {
-                        tracing::info!("Revert protection enabled");
-
-                        let pool = ctx.pool().clone();
-                        let provider = ctx.provider().clone();
-                        let revert_protection_ext = RevertProtectionExt::new(pool, provider);
-
-                        ctx.modules
-                            .merge_configured(revert_protection_ext.into_rpc())?;
-                    }
-
-                    Ok(())
-                })
-                .on_node_started(move |ctx| {
-                    version.register_version_metrics();
-                    if builder_args.log_pool_transactions {
-                        tracing::info!("Logging pool transactions");
-                        ctx.task_executor.spawn_critical(
-                            "txlogging",
-                            Box::pin(async move {
-                                monitor_tx_pool(ctx.pool.all_transactions_event_listener()).await;
-                            }),
-                        );
-                    }
-
-                    Ok(())
-                })
-                .launch()
-                .await?;
-
-            handle.node_exit_future.await
-        })
-        .unwrap();
 }
+
+/// Starts the OP builder node with a given payload builder implementation.
+fn start_builder_node<B: builders::PayloadBuilder>(cli: Cli)
+where
+    BuilderConfig<<B as builders::PayloadBuilder>::Config>: TryFrom<OpRbuilderArgs>,
+    <BuilderConfig<<B as builders::PayloadBuilder>::Config> as TryFrom<OpRbuilderArgs>>::Error:
+        Debug,
+{
+    cli.run(|builder, builder_args| async move {
+        let builder_config = BuilderConfig::<B::Config>::try_from(builder_args.clone())
+            .expect("Failed to convert rollup args to builder config");
+
+        let rollup_args = builder_args.rollup_args;
+        let op_node = OpNode::new(rollup_args.clone());
+        let handle = builder
+            .with_types::<OpNode>()
+            .with_components(
+                op_node
+                    .components()
+                    .pool(
+                        OpPoolBuilder::<FBPooledTransaction>::default()
+                            .with_enable_tx_conditional(
+                                // Revert protection uses the same internal pool logic as conditional transactions
+                                // to garbage collect transactions out of the bundle range.
+                                rollup_args.enable_tx_conditional
+                                    || builder_args.enable_revert_protection,
+                            )
+                            .with_supervisor(
+                                rollup_args.supervisor_http.clone(),
+                                rollup_args.supervisor_safety_level,
+                            ),
+                    )
+                    .payload(B::new_service(builder_config)?),
+            )
+            .with_add_ons(
+                OpAddOnsBuilder::default()
+                    .with_sequencer(rollup_args.sequencer.clone())
+                    .with_enable_tx_conditional(rollup_args.enable_tx_conditional)
+                    .build(),
+            )
+            .extend_rpc_modules(move |ctx| {
+                if builder_args.enable_revert_protection {
+                    tracing::info!("Revert protection enabled");
+
+                    let pool = ctx.pool().clone();
+                    let provider = ctx.provider().clone();
+                    let revert_protection_ext = RevertProtectionExt::new(pool, provider);
+
+                    ctx.modules
+                        .merge_configured(revert_protection_ext.into_rpc())?;
+                }
+
+                Ok(())
+            })
+            .on_node_started(move |ctx| {
+                VERSION.register_version_metrics();
+                if builder_args.log_pool_transactions {
+                    tracing::info!("Logging pool transactions");
+                    ctx.task_executor.spawn_critical(
+                        "txlogging",
+                        Box::pin(async move {
+                            monitor_tx_pool(ctx.pool.all_transactions_event_listener()).await;
+                        }),
+                    );
+                }
+
+                Ok(())
+            })
+            .launch()
+            .await?;
+
+        handle.node_exit_future.await
+    })
+    .unwrap();
+}
+
+const VERSION: VersionInfo = VersionInfo {
+    version: CARGO_PKG_VERSION,
+    build_timestamp: VERGEN_BUILD_TIMESTAMP,
+    cargo_features: VERGEN_CARGO_FEATURES,
+    git_sha: VERGEN_GIT_SHA,
+    target_triple: VERGEN_CARGO_TARGET_TRIPLE,
+    build_profile: BUILD_PROFILE_NAME,
+};
