+++ crates/op-rbuilder/src/primitives/reth/execution.rs
@@ -1,10 +1,11 @@
 //! Heavily influenced by [reth](https://github.com/paradigmxyz/reth/blob/1e965caf5fa176f244a31c0d2662ba1b590938db/crates/optimism/payload/src/builder.rs#L570)
 use alloy_consensus::Transaction;
 use alloy_primitives::{private::alloy_rlp::Encodable, Address, U256};
+use core::fmt::Debug;
 use reth_optimism_primitives::{OpReceipt, OpTransactionSigned};
 
 #[derive(Default, Debug)]
-pub struct ExecutionInfo {
+pub struct ExecutionInfo<Extra: Debug + Default = ()> {
     /// All executed transactions (unrecovered).
     pub executed_transactions: Vec<OpTransactionSigned>,
     /// The recovered senders for the executed transactions.
@@ -17,12 +18,11 @@ pub struct ExecutionInfo {
     pub cumulative_da_bytes_used: u64,
     /// Tracks fees from executed mempool transactions
     pub total_fees: U256,
-    #[cfg(feature = "flashblocks")]
-    /// Index of the last consumed flashblock
-    pub last_flashblock_index: usize,
+    /// Extra execution information that can be attached by individual builders.
+    pub extra: Extra,
 }
 
-impl ExecutionInfo {
+impl<T: Debug + Default> ExecutionInfo<T> {
     /// Create a new instance with allocated slots.
     pub fn with_capacity(capacity: usize) -> Self {
         Self {
@@ -32,8 +32,7 @@ impl ExecutionInfo {
             cumulative_gas_used: 0,
             cumulative_da_bytes_used: 0,
             total_fees: U256::ZERO,
-            #[cfg(feature = "flashblocks")]
-            last_flashblock_index: 0,
+            extra: Default::default(),
         }
     }
